refactor: move business logic of ladder game to ladder class
